ai_service/requirements.txt
fastapi==0.110.0
uvicorn==0.27.1
numpy==1.26.4
pillow==10.3.0
opencv-python==4.9.0.80
mediapipe==0.10.14
